+++ api/src/controllers/authController.js
@@ -2,12 +2,10 @@ const { findAllUsers, findUserById, findDriverById } = require("../Models/utils/
 const bcrypt = require("bcrypt");
 const {
   isCorrectCredentials,
-  isFitToBuy,
 } = require("../Models/utils/Confirmer");
 const {
   createUser,
   createDriver,
-  assingTrip,
 } = require("../Models/utils/Creations");
 
 const jwt = require("jsonwebtoken");
@@ -128,24 +126,6 @@ const getDriverById = async (req, res) => {
   }
 }
 
-const purchaseTrip = async (req, res) => {
-  try {
-    const { user_id, trip_id, capacity} = req.body;
-    if (!user_id || !trip_id)
-      throw new Error("Faltan datos del viaje o del usuario");
-    const canBuy = await isFitToBuy(user_id, trip_id);
-    if (!canBuy)
-      return res
-        .status(401)
-        .json({ error: "No estas autorizado para comprar este viaje" });
-    //vincular viaje
-    const trip = await assingTrip({ user_id, trip_id, capacity });
-    res.json({ trip_purchased: trip });
-  } catch (e) {
-    res.status(400).json({ error: `${e.message}` });
-  }
-};
-
 module.exports = {
   registerUser,
   getUsers,
@@ -153,7 +133,6 @@ module.exports = {
   getDriverById,
   registerDriver,
   loginUser,
-  purchaseTrip,
   verifyUser,
   verifyBanStatus,
 };
